requirements.txt
beautifulsoup4==4.13.3
playwright==1.50.0
python-dotenv==1.0.1
requests==2.31.0
supabase==2.13.0